More fixes to autofisher, add FISH to allowed_actions
The FISH action is used on things like waterygrave, flotsam, etc. where
the action works directly, rather than through the fish -> hook -> reel
cycle as with ponds. The changes committed should make sure that
autofisher is only used when the action cycle is needed, and otherwise
the normal fish action in the normal repeat cycle is used.