Fix Wilba crash, drop my DebugPrints
In were-form, Wilba's player ent overrides the actionpicker component's
GetClickAction. The overriding function does not always provide a
non-empty list of functions - the Werewilba is unable to examine
in-world items (despite being able to do this with items in her
inventory).
When this happens, just skip over the iteration, which doesn't handle
nil actionlist.
Also, a ton of DebugPrints added by me were removed, it seems the
existing features work as well as they do in DST.