Fix wolfgang bug
If wolfgang's powerdown transformation hits while we're in dolongaction state,
the action is interrupted, but the bufferedaction is left set. This, because
of how playercontroller handles DoAction calls, prevents this action from
being ever carried out. Detect this by checking if we are in idle state and
also have bufferedaction set, and clear the buffered action from AQR code..